requirements.txt
dj-database-url==0.4.1
Django==1.10.2
django-bootstrap3==7.1.0
djangorestframework==3.4.7
gunicorn==19.6.0
Markdown==2.6.7
Pillow==3.4.2
psycopg2==2.6.2
whitenoise==3.2.2